Fix warnings after enabling -Wextra
Almost all of the warnings are related to signed vs unsigned
comparisons.

Key changes:
 * use time_t instead of uint64_t for variables that represent
   a unix timestamp (including intercept start and end times in
   intercept_common_t)
 * replace loop iterator variables with size_t where appropriate

+++ src/collector/email_ingest_service.c
@@ -168,7 +168,7 @@ static MHD_RESULT iterate_post (void *coninfo_cls, enum MHD_ValueKind kind,
                 ptr ++;
             }
 
-            if (*ptr == '\0' || ptr - data >= size) {
+            if (*ptr == '\0' || (size_t)(ptr - data) >= size) {
                 free(con_info->thismsg->content);
                 con_info->thismsg->content = NULL;
             }
